+++ No.h
@@ -19,12 +19,12 @@ class No{
         Aresta* ultimaAresta;
 
         int demanda;
-        int x;
-        int y;
+        double x;
+        double y;
 
     public:
-        No(int idNo, int x, int y, int demanda);
-        No(int idNo, int x, int y);
+        No(int idNo, double x, double y, int demanda);
+        No(int idNo, double x, double y);
         No(int idNo, int pesoNo);
         No(int idNo);
         ~No();
@@ -41,8 +41,8 @@ class No{
         Aresta* getAresta(int idNoDestino);
 
         int getDemanda();
-        int getX();
-        int getY();
+        double getX();
+        double getY();
 
         //SETTERS
         void setProxNo(No* no);
@@ -50,8 +50,8 @@ class No{
         void setVisita(bool visitadoOuNao);
 
         void setDemanda(int demanda);
-        void setX(int x);
-        void setY(int y);
+        void setX(double x);
+        void setY(double y);
 
         //OUTROS METODOS
         bool temAresta(int idNoDestino);
